# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Poultry Trading ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, agent sees own)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithDetails"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (agent or admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
  /orders/{id}/history:
    get:
      operationId: getOrderHistory
      tags: [orders]
      summary: Get order status change history
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderHistoryEntry"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        username:
          type: string
        role:
          type: string
          enum: [admin, agent]
        createdAt:
          type: string
      required:
        - id
        - name
        - username
        - role
        - createdAt
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        openingBalance:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - openingBalance
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [chicks, chickens, other]
        unitPrice:
          type: string
        stockQuantity:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - unitPrice
        - stockQuantity
        - createdAt
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: string
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - orderId
        - productId
        - quantity
        - unitPrice
    OrderWithDetails:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        agentId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, preparing, delivering, delivered, cancelled]
        orderDate:
          type: string
        deliveryDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        agent:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - customerId
        - status
        - orderDate
        - createdAt
        - customer
        - items
    CreateOrderRequest:
      type: object
      properties:
        customerId:
          type: integer
        agentId:
          type: integer
        orderDate:
          type: string
        deliveryDate:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerId
        - orderDate
        - items
    UpdateOrderRequest:
      type: object
      properties:
        customerId:
          type: integer
        agentId:
          type: integer
        status:
          type: string
          enum: [pending, preparing, delivering, delivered, cancelled]
        orderDate:
          type: string
        deliveryDate:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, delivering, delivered, cancelled]
      required:
        - status
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: string
      required:
        - productId
        - quantity
        - unitPrice
    OrderHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        changedBy:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        oldStatus:
          $ref: "#/components/schemas/OrderWithDetailsStatus"
        newStatus:
          $ref: "#/components/schemas/OrderWithDetailsStatus"
        changedAt:
          type: string
      required:
        - id
        - orderId
        - changedBy
        - oldStatus
        - newStatus
        - changedAt
...