# lib/api-zod/src/generated/types/orderHistoryEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry Trading ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderHistoryEntryChangedBy } from "./orderHistoryEntryChangedBy";
import type { OrderWithDetailsStatus } from "./orderWithDetailsStatus";

export interface OrderHistoryEntry {
  id: number;
  orderId: number;
  changedBy: OrderHistoryEntryChangedBy;
  oldStatus: OrderWithDetailsStatus;
  newStatus: OrderWithDetailsStatus;
  changedAt: string;
}